feature(profile): change navbar options depending on logged in user
If the logged in user is viewing their own profile page, the navbar
options will display all of the possible options. Otherwise, when
viewing another user's profile page or viewing a profile page when
not logged in, the navbar options will just be the default public
options: "Book Listings" and "Reviews".

Also styled the Profile page in proper page format.